+++ changelog.md
@@ -5,6 +5,12 @@
 ### Added
 
 - Added new field type for geo coordinates to output them with 4 digits
+- Added internal names for each page builder support
+- Added 9 abilities to get object data or start imports via AI
+
+### Changed
+
+- Use $throwable::class instead of get_class() for the modern way of PHP
 
 ### Fixed